Implement Event Counter Please write a small library which helps to track the number of events that happened during a specified window of time. As an example, it might be used to track how often a particular webpage is served.
The library should support two operations:
Client should be able to signal that a single event happened (e.g. a page was served)
Client should be able to request the number of events that happened over a user-specified amount of time until current time. You may limit the supported timespan to a reasonable upper bound, like 5 minutes.

Please only track counts, you do not have to track any data payload associated with an event.

Since the main use for a component like this is presenting diagnostic information on demand, it doesn’t have to persist data. Additionally, please do not spend time writing an example application that integrates the library, because we will not base our assessment on it.